支持hq协议

+++ xquic/src/main/cpp/main/src/xquic_h3_callbacks.c
@@ -233,15 +233,15 @@ int client_h3_request_read_notify(xqc_h3_request_t *h3_request, xqc_request_noti
                                 user_stream->user_tag);
 
         /* auto to close request */
-        int ret = xqc_h3_request_finish(h3_request);
+        ssize_t ret = xqc_h3_request_finish(h3_request);
         //int ret = xqc_h3_request_close(h3_request);
-        LOGD("auto to call xqc_h3_request_close ret=%d", ret);
+        LOGD("auto to call xqc_h3_request_close ret=%zd", ret);
 
         if (user_stream->user_conn->ctx->args->net_cfg.conn_type == CONN_TYPE_SHORT) {
             /* auto to close conn */
             ret = xqc_h3_conn_close(user_stream->user_conn->ctx->engine,
                                     &user_stream->user_conn->cid);
-            LOGD("auto to call xqc_h3_conn_close ret=%d", ret);
+            LOGD("auto to call xqc_h3_conn_close ret=%zd", ret);
         }
     }
     return 0;
